Update README.md
We began as a modest high-street store in the UK that sold printing supplies to both business and retail clients. However, a lot has changed since then!
We immediately recognized the potential when a friend told us about a “vaping device” they were using to stop smoking.
We threw ourselves into this new and exciting market because we wanted to make a difference.
We soon started offering e-liquids and simple vaping kits, which attracted a lot of interest right away.
We developed the “Geek Bar Vape Shop UK” brand in 2014, and it was incorporated into our store.  https://insurancegeekbar.com/product/geek-bar-banana-vanilla-online-store/ https://insurancegeekbar.com/product/geek-bar-banana-vanilla-online-store/